backend/pyproject.toml
[project]
name = "haulage-truck-management"
version = "0.1.0"
description = "Haulage Truck Management System API"
requires-python = ">=3.11"

dependencies = [
    "Django>=5.0",
    "djangorestframework>=3.15",
    "django-filter>=24.2",
    "python-dotenv>=1.0",
    "gunicorn>=21.2",
    "djangorestframework-simplejwt>=5.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "factory-boy>=3.3",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.development"
python_files = ["tests.py", "test_*.py"]